Fix issue where Delete Post would fail

+++ app/controllers/posts_controller.rb
@@ -46,11 +46,16 @@ class PostsController < ApplicationController
 
   # The destroy action removes the post permanently from the database
   def destroy
-    if @post.destroy
-      flash[:notice] = "Successfully deleted post!"
-      redirect_to posts_path
+    @post = Post.find(params[:id])
+    if @post.present?
+      if @post.destroy
+        flash[:notice] = "Successfully deleted post!"
+        redirect_to posts_path
+      else
+        flash[:alert] = "Error deleting post!"
+      end
     else
-      flash[:alert] = "Error updating post!"
+      flash[:alert] = "Error deleting post!"
     end
   end
